requirements.txt
beautifulsoup4
matplotlib
numpy
openai
pandas
Pillow
pymupdf
pyvis
python-dotenv
reportlab
requests
seaborn
streamlit